+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## next
 
+- Update to [Neucore plugin](https://github.com/tkhamez/neucore-plugin) version 2.1.0:
+  - Menu items can now require the user to be a member or a manager of a group.
 - Fix: Menu items from plugins were not always shown before login.
 - Text changes.